+++ src/pg_numhstore.h
@@ -12,7 +12,7 @@
 
 HStore * hstorePairs( Pairs *pairs, int4 pcount, int4 buflen );
 HStore * hstoreUpgrade( Datum orig );
-int adeven_add_get_digit_num( long number );
+size_t adeven_add_get_digit_num( long number );
 extern void get_typlenbyvalalign( Oid eltype, int16 * i_typlen, bool * i_typbyval, char * i_typalign );
 
 #endif
